move close notebook to server state handler
